fix(week-calendar): address branch-review findings (blockers + non-blockers)
Backend
- H1: wire the schema migration into run_migrations() — widen the 4 hours
  columns to NUMERIC + add start_time/end_time + index on existing DBs
  (idempotent), so prod Postgres isn't left INTEGER/missing columns.
- H2/M2: flip remaining int hours schemas to float (admin EmployeeTicketResponse,
  WeeklyTicket, TimeEntryRow.hours, TimeEntriesResponse.total_hours; workitems
  MyTaskResponse.assigned_hours) — they would 500 on fractional values.
- H3/L3: coerce inbound datetimes to naive-UTC at the API boundary; explicit
  None guards in update (no truthy `or`).
- M3: narrow the week query to the 5 rendered days (Mon–Fri).
- M6/L1: note ALTER-TYPE table-rewrite cost; stop int() truncation in the
  dedupe maintenance script.
- M5: add tests (delete authz, reassign-to-unowned, null-position PATCH,
  aware-datetime normalization, 5-day window, inclusive/exclusive boundary).
- M1: contract goldens regenerated with a fractional fixture so the harness
  actually exercises the int->float contract.

Frontend
- H4: ignore drag on optimistic/draft (id<=0) blocks.
- H5: TimeBlock keyboard-selectable (Enter/Space) so the arrow/Delete model is reachable.
- M4: memoize drag callbacks (listeners bind once); fix the false
  setPointerCapture comment + brittle draw-dayIdx ternary.
- M7/L4: loading + error states on the initial fetch; now-line advances on a 60s tick.
- L2: hoist calendar accent/capacity colors into lib/calendarTheme.ts.

Docs: REVIEW_RULES.md capturing the fractional-hours / rollup / authz / naive-UTC
/ migration / golden-harness invariants.

+++ backend/routers/workitems.py
@@ -634,7 +634,7 @@ class MyTaskResponse(BaseModel):
     is_overdue: bool
     completed_at: str | None = None
     story_points: int = 0
-    assigned_hours: int = 0
+    assigned_hours: float = 0
     assignee: str
     reporter_name: str | None = None
     description: str = ""
